apiRecommend/requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas
sqlalchemy
apscheduler
python-dotenv